Remove unused logger
